refactor tree view
