+++ docs/src/pages/tools/digital-twin-designer.mdx
@@ -1,0 +1,11 @@
+---
+order: 9
+title: Digital Twin Designer
+---
+
+import DigitalTwinDesigner from "../../components/DigitalTwinDesigner"
+
+A [digital twin definition](https://github.com/Azure/opendigitaltwins-dtdl/) designer to be used in IoT solutions such as with Azure IoT Hubs 
+or Azure Digital Twins
+
+<DigitalTwinDesigner />
